oh I haven't done this in a while - the game works now
Added enemy spawning and a class which takes a delegate from the hud so spawnable objects can update the hud when they collide with the player. used by the pickups, enemies and hazards. You can set the effect it has on the game from the editor, its a serialized enum and it corresponds with the function that'll eventually be called in the hud.  this worked out great, i've never been prouder of a piece of code.